refactor: move implementation into Client module
Main module delegates all of its behavior to the Client. The prepares us to add a supervisor at the top level.
